Remove imports under TYPE_CHECKING and resolve circular dependency
