+++ src/botamusique/main.py
@@ -12,11 +12,12 @@ from typing import Any
 
 from botamusique import command
 from botamusique import constants
+from botamusique import interface
 from botamusique import util
 from botamusique.database import SettingsDatabase, MusicDatabase, DatabaseMigration
 from botamusique.media.cache import MusicCache
 from botamusique.media.playlist import get_playlist
-from botamusique.mumbleBot import MumbleBot, start_web_interface
+from botamusique.mumbleBot import MumbleBot
 
 
 def main() -> None:
@@ -266,6 +267,29 @@ def main() -> None:
 
     # Start the main loop.
     bot.loop()
+
+def start_web_interface(addr: str, port: int, bot: MumbleBot) -> None:
+    # setup logger
+    werkzeug_logger = logging.getLogger('werkzeug')
+    werkzeug_logger.propagate = False
+    logfile = util.solve_filepath(bot.config.get('webinterface', 'web_logfile'))
+    if logfile:
+        handler = logging.handlers.RotatingFileHandler(logfile, mode='a', maxBytes=10240, backupCount=3)  # Rotate after 10KB, leave 3 old logs
+    else:
+        handler = logging.StreamHandler()
+
+    # replace werkzeug_logger handlers with ours
+    for handler in list(werkzeug_logger.handlers):
+        if isinstance(handler, logging.StreamHandler):
+            werkzeug_logger.removeHandler(handler)
+            handler.close()
+    werkzeug_logger.addHandler(handler)
+
+    interface.init_app()
+    interface.set_bot(bot)
+    interface.init_proxy()
+    interface.web.secret_key = bot.config.get('webinterface', 'flask_secret')
+    interface.web.run(port=port, host=addr)
 
 if __name__ == '__main__':
     main()
